수정. taskMain 리팩토링
